refactor: reorganize source code into feature modules
