+++ viberwhisper/src/audio/mod.rs
@@ -1,0 +1,2 @@
+pub mod recorder;
+pub use recorder::AudioRecorder;
