+++ viberwhisper/src/input/mod.rs
@@ -1,0 +1,3 @@
+pub mod hotkey;
+pub mod tray;
+pub mod typer;
