+++ viberwhisper/src/platform/mod.rs
@@ -1,0 +1,4 @@
+#[cfg(target_os = "macos")]
+pub mod macos;
+#[cfg(target_os = "windows")]
+pub mod windows;
