+++ viberwhisper/src/transcriber/mod.rs
@@ -1,0 +1,2 @@
+pub mod groq;
+pub use groq::{GroqTranscriber, MockTranscriber, Transcriber};
